chore: move runtime tests into `runtime_tests`
