--- jest.lambda.config.js
@@ -1,6 +1,0 @@
-export default {
-  testMatch: ['**/test_lambda/**/*.+(ts|tsx|js)'],
-  transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-  },
-}
