--- jest.wrangler.config.js
@@ -1,6 +1,0 @@
-export default {
-  testMatch: ['**/test_wrangler/**/(*.)+(test).+(ts|tsx)'],
-  transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
-  },
-}
