+++ package.json
@@ -11,13 +11,13 @@
   ],
   "scripts": {
     "test": "jest",
-    "test:deno": "env NAME=Deno deno test --allow-read --allow-env test_deno",
-    "test:bun": "env NAME=Bun bun test --jsx-import-source ../src/middleware/jsx/jsx-dev-runtime test_bun/index.test.tsx",
-    "test:fastly": "jest --config ./jest.fastly.config.js",
-    "test:lagon": "start-server-and-test \"lagon dev test_lagon/index.ts\" http://127.0.0.1:1234 \"yarn jest test_lagon/index.test.ts --testMatch '**/*.test.ts'\"",
-    "test:node": "env NAME=Node jest --config ./jest.node.config.js",
-    "test:wrangler": "jest --config ./jest.wrangler.config.js",
-    "test:lambda": "env NAME=Node jest --config ./jest.lambda.config.js",
+    "test:deno": "env NAME=Deno deno test --allow-read --allow-env runtime_tests/deno",
+    "test:bun": "env NAME=Bun bun test --jsx-import-source ../../src/middleware/jsx/jsx-dev-runtime runtime_tests/bun/index.test.tsx",
+    "test:fastly": "jest --config ./runtime_tests/fastly/jest.config.js",
+    "test:lagon": "start-server-and-test \"lagon dev runtime_tests/lagon/index.ts\" http://127.0.0.1:1234 \"yarn jest runtime_tests/lagon/index.test.ts --testMatch '**/*.test.ts'\"",
+    "test:node": "env NAME=Node jest --config ./runtime_tests/node/jest.config.js",
+    "test:wrangler": "jest --config ./runtime_tests/wrangler/jest.config.js",
+    "test:lambda": "env NAME=Node jest --config ./runtime_tests/lambda/jest.config.js",
     "test:all": "yarn test && yarn test:deno && yarn test:bun && yarn test:fastly && yarn test:lagon && yarn test:node && yarn test:wrangler && yarn test:lambda",
     "lint": "eslint --ext js,ts src .eslintrc.cjs",
     "lint:fix": "eslint --ext js,ts src .eslintrc.cjs --fix",
