+++ runtime_tests/bun/favicon.ico
