+++ runtime_tests/fastly/jest.config.js
@@ -1,7 +1,8 @@
 export default {
-  testMatch: ['**/test_fastly/**/*.+(ts|tsx|js)'],
+  testMatch: ['**/runtime_tests/fastly/**/*.+(ts|tsx|js)'],
   transform: {
     '^.+\\.(ts|tsx)$': 'ts-jest',
   },
   preset: 'jest-preset-fastly-js-compute/typescript/esm',
+  testPathIgnorePatterns: ['jest.config.js'],
 }
