+++ runtime_tests/lambda/jest.config.js
@@ -1,0 +1,7 @@
+export default {
+  testMatch: ['**/runtime_tests/lambda/**/*.+(ts|tsx|js)'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  testPathIgnorePatterns: ['jest.config.js'],
+}
