+++ runtime_tests/node/jest.config.js
@@ -1,0 +1,7 @@
+export default {
+  testMatch: ['**/runtime_tests/node/**/*.+(ts|tsx|js)'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  testPathIgnorePatterns: ['jest.config.js'],
+}
