+++ runtime_tests/wrangler/jest.config.js
@@ -1,0 +1,7 @@
+export default {
+  testMatch: ['**/runtime_tests/wrangler/**/(*.)+(test).+(ts|tsx)'],
+  transform: {
+    '^.+\\.(ts|tsx)$': 'ts-jest',
+  },
+  testPathIgnorePatterns: ['jest.config.js'],
+}
